build write withdraw skeleton
